ovn-northd: Clear SB records depending on stale datapaths.
In case ovn-northd connects to a Southbound database that is inconsistent
with the current Northbound database there might be stale records in the
SB Datapath_Binding table. These records should be removed from the
database and all referring records from other tables should be
removed too:
- Mac_Binding
- Logical_Flow
- Multicast_Group
- IP_Multicast
- IGMP_Group

Two potential scenarios that might trigger this issue are:
a. ovn-northd is started and pointed to instances of NB and SB databases
   that have inconsistent contents.
b. NB/SB databases are running in clustered mode and bugs in ovsdb-server
   cause the NB/SB instance ovn-northd reconnects to to be in an inconsistent
   state.

Reported-by: Dan Williams <dcbw@redhat.com>
Reported-at: https://bugzilla.redhat.com/1828637
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>