ovn-northd: Remove Port_Binding stale references to HA_Chassis_Group.
If a logical switch port of type "external" that has ha_chassis_group set
is deleted from a logical switch and a port with the same name is added
to a different logical switch in the same transaction, the SB Port_Binding
record corresponding to the logical port is reused but the ha_chassis_group
field is not cleared.

One way to reproduce the issue is:
$ ovn-nbctl ls-add ls1
$ ovn-nbctl ls-add ls2
$ ovn-nbctl lsp-add ls1 lsp1
$ ovn-nbctl lsp-add ls2 lsp2
$ ovn-nbctl lsp-set-type lsp2 external
$ ovn-nbctl ha-chassis-group-add chg1
$ ovn-nbctl ha-chassis-group-add-chassis chg1 chassis-1 30
$ chg1_uuid=$(ovn-nbctl --bare --columns _uuid list ha_Chassis_Group .)
$ ovn-nbctl set logical_switch_port lsp2 ha_chassis_group=${chg1_uuid}
$ ovn-nbctl lsp-del lsp2 -- lsp-add ls1 lsp2

To avoid that, we now clear the Port_Binding ha_chassis_group field when
the logical port corresponding to the port binding is not of type
"external" anymore.

Reported-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>