Support packet metadata marking for logical port traffic.
This patch adds a new column 'mark' in the Logical_Switch_Port
table in the NB DB. CMS can set a desired value in this column
and whenever a packet is received from the VIF of the logical port,
the NXM_NX_PKT_MARK field is set with the value specified by CMS.

In the case of Linux, this corresponds to struct sk_buff's "skb_mark"
member and this mark can be seen by the linux networking subsystem.
CMS can inspect this value (as an iptables rule or adding an OF flow
in another ovs bridge) and take appropriate action.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>