ovn-controller to no longer monitor Chassis' external_ids
Prior to this patch, the external_ids column from the Chassis table were
being used for two purposes:

1. Holding configuration for the OVN (copied by ovn-controller from the
local OVS database)

2. Holding information from external systems (the main purpose of the
   external_ids column across other resources).

The problem with mixing these two use cases is that, ovn-controller
should be aware of changes to the configuration and it would trigger
flow re-computations upon those changes. It shouldn't care tho about the
information from external systems but, since these two things were put
together, CMSs writing things to the external_ids column of the Chassis
were waking up ovn-controller to recompute flows.

This patch is separating these two things by creating another column in
the Chassis table called "other_config". This new table holds the OVN
configuration that is copied over from the local OVS db leaving the
external_ids column unmonitored and free for other systems to make use
of it.

This change also keeps things backward compatible by continuing to
write the OVN configuration to the external_ids column for external
systems that may be reading them from there but, that column is no
longer monitored so it won't generate any events. This behavior should
be temporary and removed in the future. The note in the NEWS file and
comments in the code itself points to the future removal of this
behavior.

Reported-At: https://bugzilla.redhat.com/show_bug.cgi?id=1824220
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>