Don't leak values other than 1 or 0 as bool return values on C89 compiler
While the code base makes use of bool type defined in C99, we don't assume C99
semantics for integer or pointer conversions to bool. (This is explained in the
project coding style guide.) In C89 environments, it's important to normalize
bool variables and returned values to 1 | 0 to avoid issues down the line.

This patch updates all functions in the tree that return bool values and that
could, in degenerate cases, return bools set to values different from 1 | 0.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>