Add expression writeability scopes.
Logical fields are defined as either being writeable or read-only. There
is no way to make fields writeable only in specific scenarios.

This commit changes the boolean writeability field to a field of flags
indicating contexts where a field is writeable. Any time that nested
actions are used (i.e. actions enclosed in curly braces), a new scope
may be set for the nested action. For this particular commit, no
functionality is changed, and only a "default" scope is added
that mirrors the current setup. A future commit will make use of this
feature.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>