Used nested actions in ct_commit
ct_commit allows for ct_label and ct_mark to be set within. However,
there are some restrictions with the current implementation:

* It is not possible to address the indiviual bits within the ct_mark or
   ct_label.
* It is not possible to set these to the value of a register. Only
   explicit integer setting can be used.

With this change, ct_commit now can have arbitrary nested actions
inside. This makes it similar to how the "exec" option works in OVS's
ct() action.

ct_commit now also sets a writeability scope so that ct_mark and
ct_label are the only symbols that are writeable. The positive side
effect is that ct_mark and ct_label are no longer writeable except for
inside ct_commit.

In this commit, the only noticeable effect is that it allows for
slightly more expressive setting of ct_label.blocked. A future commit
will take further advantage of this.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>