pinctrl: Avoid flushing of non-local IGMP_Groups.
With ovn-monitor-all enabled, all ovn-controllers get updates about all
SB records. This means that an ovn-controller might receive updates about
IGMP_Groups inserted by other chassis on datapaths that are not local to
the local chassis. These entries are valid and are not owned by the local
chassis and should not be flushed.

Fixes: 70ff8243040f ("OVN: Add IGMP SB definitions and ovn-controller support")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>