ovn-northd: By default don't enable probe for unix socket.
The commit ed09854e5b3 ("ovn-northd: Add the option to configure probe
interval") introduced the option northd-probe-interval to configure
ovn-northd probe interval with NB/SB DBs. However, it overrided the
default behavior that it used before - setting default probe interval
according to stream type. E.g., for unix socket, probe was disabled by default
but with the change it defaults to 5 seconds, if the option
northd-probe-interval is not set.

This patch restores the default behavior, while honoring the settings of
northd-probe-interval when it is configured.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>