Provide timestamp information for nb_cfg.
This change adds timestamp information for nb_cfg related processing of
ovn-northd and ovn-controller. It is useful for measuring end to end
control plane latency more accurately. The timestamps tell exactly when
ovn-northd starts processing a NB DB change, updates the changes to
SB DB, and when the slowest chassis finishes processing the changes.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>