ovn-nbctl: Support option --print-wait-time for latency measurement.
This patch adds option --print-wait-time, which can be used together
with --wait=sb/hv to print the time spent on respective processing.

For example:
$ ovn-nbctl --print-wait-time --wait=hv sync
Time spent on processing nb_cfg 392:
    ovn-northd delay before processing: 1ms
    ovn-northd completion:              32ms
    ovn-controller(s) completion:       39ms

Without this, we had to use "time" command to measure the total time
spent by the ovn-nbctl command. However, it doesn't refect the actual
control plane latency, because the nb_cfg updates from all hypervisors
to SB DB contributes a significant part to the time spent on waiting
by the ovn-nbctl command. This doesn't reflect the real production
scenario. With this option (and the timestamp support from the previous
commit), we can measure the exact time when hypervisors completed
SB DB change handling and enforcing to OVS.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>