Add 2020 and 2021 release calendars.
In 2020, there were times where we started the release process later
than expected. Part of the reason why was that the actual release dates
for OVN were vague, meaning nobody really knew when we should start the
release process.

This attempts to remedy the problem by creating a calendar for releases
for 2020 and 2021. This way developers can know more easily when
releases are scheduled and prepare in advance better.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>