ovn-controller: Unix command debug/delay-nb-cfg-report.
This command is added to avoid the flooding of nb_cfg updates from
a large nubmer of hypervisors interfering with the original SB DB
data distribution and handling, so that the e2e control plane latency
can be measured more accurately during scale testing with the command
    ovn-nbctl --print-wait-time --wait=hv ...

Without this, there can be hypervisors updating back to SB before the
SB server finishes sending notifications to the rest of hypervisors.
In my test with 3K HVs, delaying the report for 2 seconds is enough to
get accurate latency report.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>